requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.4
Babel==2.8.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.4
cryptography==3.1
cycler==0.10.0
dnspython==2.0.0
fpdf==1.7.2
idna==2.10
isort==5.6.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
matplotlib==3.3.0
mccabe==0.6.1
minecart==0.3.0
MouseInfo==0.1.3
multidict==4.7.6
numpy==1.19.1
pdf2image==1.14.0
pdfminer3k==1.3.4
Pillow==7.2.0
ply==3.11
prettytable==0.7.2
psycopg2==2.8.5
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.9
pylint==2.6.0
pymongo==3.10.1
PyMsgBox==1.0.9
PyMySQL==0.10.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.25.10
wrapt==1.12.1
yarl==1.4.2
